move LabelValueList to lib-components
