+++ frontend/src/employee-frontend/components/child-information/assistance/AssistanceActionRow.tsx
@@ -22,13 +22,13 @@ import { useMutationResult } from 'lib-common/query'
 import { scrollToRef } from 'lib-common/utils/scrolling'
 import UnorderedList from 'lib-components/atoms/UnorderedList'
 import { FixedSpaceColumn } from 'lib-components/layout/flex-helpers'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { LabelLike } from 'lib-components/typography'
 import { featureFlags } from 'lib-customizations/employee'
 
 import { useTranslation } from '../../../state/i18n'
 import { UIContext } from '../../../state/ui'
 import { isActiveDateRange } from '../../../utils/date'
-import LabelValueList from '../../common/LabelValueList'
 import Toolbar from '../../common/Toolbar'
 import ToolbarAccordion from '../../common/ToolbarAccordion'
 import { deleteAssistanceActionMutation } from '../queries'
