+++ frontend/src/employee-frontend/components/child-information/person-details/AdditionalInformation.tsx
@@ -22,6 +22,7 @@ import Combobox from 'lib-components/atoms/dropdowns/Combobox'
 import Checkbox from 'lib-components/atoms/form/Checkbox'
 import TextArea from 'lib-components/atoms/form/TextArea'
 import { FixedSpaceRow } from 'lib-components/layout/flex-helpers'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { H4 } from 'lib-components/typography'
 import { featureFlags } from 'lib-customizations/employee'
 import { faPen } from 'lib-icons'
@@ -32,7 +33,6 @@ import { UIContext } from '../../../state/ui'
 import { formatParagraphs } from '../../../utils/html-utils'
 import { RequireRole } from '../../../utils/roles'
 import { renderResult } from '../../async-rendering'
-import LabelValueList from '../../common/LabelValueList'
 import { FlexRow } from '../../common/styled/containers'
 import { textAreaRows } from '../../utils'
 import {
