+++ frontend/src/employee-frontend/components/fee-decision-details/ChildSection.tsx
@@ -12,11 +12,11 @@ import type {
 import type { PlacementType } from 'lib-common/generated/api-types/placement'
 import { formatPersonName } from 'lib-common/names'
 import CollapsibleSection from 'lib-components/molecules/CollapsibleSection'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { PersonName } from 'lib-components/molecules/PersonNames'
 import { faUserFriends } from 'lib-icons'
 
 import { useTranslation } from '../../state/i18n'
-import LabelValueList from '../common/LabelValueList'
 
 interface Props {
   child: PersonDetailed
