+++ frontend/src/employee-frontend/components/invoice/InvoiceDetailsSection.tsx
@@ -7,11 +7,11 @@ import { Link } from 'wouter'
 
 import type { InvoiceDetailed } from 'lib-common/generated/api-types/invoicing'
 import CollapsibleSection from 'lib-components/molecules/CollapsibleSection'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { featureFlags } from 'lib-customizations/employee'
 import { faMoneyCheck } from 'lib-icons'
 
 import { useTranslation } from '../../state/i18n'
-import LabelValueList from '../common/LabelValueList'
 
 import { formatInvoicePeriod } from './utils'
