+++ frontend/src/employee-frontend/components/invoice/InvoiceHeadOfFamilySection.tsx
@@ -7,11 +7,11 @@ import { Link } from 'wouter'
 
 import type { InvoiceDetailed } from 'lib-common/generated/api-types/invoicing'
 import CollapsibleSection from 'lib-components/molecules/CollapsibleSection'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { PersonName } from 'lib-components/molecules/PersonNames'
 import { faUserFriends } from 'lib-icons'
 
 import { useTranslation } from '../../state/i18n'
-import LabelValueList from '../common/LabelValueList'
 
 interface Props {
   headOfFamily: InvoiceDetailed['headOfFamily']
