+++ frontend/src/employee-frontend/components/person-profile/PersonFamilyOverview.tsx
@@ -18,11 +18,11 @@ import { formatCents } from 'lib-common/money'
 import { useQueryResult } from 'lib-common/query'
 import { getAge } from 'lib-common/utils/local-date'
 import { Table, Tbody, Td, Th, Thead, Tr } from 'lib-components/layout/Table'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { PersonName } from 'lib-components/molecules/PersonNames'
 
 import { useTranslation } from '../../state/i18n'
 import { renderResult } from '../async-rendering'
-import LabelValueList from '../common/LabelValueList'
 
 import { familyByPersonQuery } from './queries'
