+++ frontend/src/employee-frontend/components/voucher-value-decision/VoucherValueDecisionHeading.tsx
@@ -11,6 +11,7 @@ import type {
   VoucherValueDecisionType
 } from 'lib-common/generated/api-types/invoicing'
 import { formatPersonName } from 'lib-common/names'
+import LabelValueList from 'lib-components/molecules/LabelValueList'
 import { PersonName } from 'lib-components/molecules/PersonNames'
 import { H1 } from 'lib-components/typography'
 import { Gap } from 'lib-components/white-space'
@@ -18,7 +19,6 @@ import colors from 'lib-customizations/common'
 
 import { getVoucherValueDecisionPdf } from '../../generated/api-clients/invoicing'
 import { useTranslation } from '../../state/i18n'
-import LabelValueList from '../common/LabelValueList'
 import WarningLabel from '../common/WarningLabel'
 import { TypeSelect } from '../fee-decision-details/TypeSelect'
