+++ frontend/src/lib-components/molecules/LabelValueList.tsx
